Changed so that the save method takes a uin16 array. A better standard for throwing information around. Will no doubt have to change something when it comes to storing grey scale

+++ Code/The_Columbus/ColumbusTest/src/CustomDevices/ImageProcessor.h
@@ -15,7 +15,7 @@
 #define BMP_FORMAT_1xUINT			4
 #define BMP_FORMAT_2xUINT8T			5
 typedef struct {
-	int *ImagePtr;
+	uint16_t *ImagePtr;
 	int Height;
 	int Width;
 	uint8_t Format;
