+++ Code/The_Columbus/ColumbusTest/src/CustomDevices/SD_Card.h
@@ -27,7 +27,7 @@ void Log_Write(char *buff, int length);
 void SaveBuff( int * WorkingBuffer , int size);
 int Read2DSignal( int * WorkingBuffer );
 int ReadSignal( int * WorkingBuffer );
-void SaveBitmap(uint8_t *Image, int width, int height, char *FileName);
+void SaveBitmap(uint16_t *Image, int width, int height, char *FileName);
 //void ReadBitmap(char *Filename);
 void ReadBitmap(char *Filename, Image_t *image);
 #endif /* SD_CARD_H_ */
